vector_service/pyproject.toml
[project]
name = "vector-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "matplotlib",
    "numpy>=2.2.6",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.3.0",
    "umap-learn",
    "uvicorn>=0.42.0",
]

